Collapse "get-weather-tool".
