+++ packages/a2/src/tools/get-weather.ts
@@ -8,133 +8,61 @@ import {
   Outcome,
   Schema,
 } from "@breadboard-ai/types";
-import { Template } from "../a2/template";
-import { defaultLLMContent, err, ok, toLLMContent, toText } from "../a2/utils";
-import getWeather, {
-  describe as getWeatherDescriber,
-  type WeatherInputs,
-} from "./get-weather-tool";
+import { err, ok, toLLMContent } from "../a2/utils";
+import { executeTool } from "../a2/step-executor";
 
 export { invoke as default, describe };
 
-/**
- * Every tool operates in one of three modes:
- * - `tool` -- when invoked as a tool by an LLM
- * - `step` -- when used as a step in a visual editor
- * - `invoke` -- when invoked directly as a module (by another module)
- */
-export type ToolMode = "tool" | "step" | "invoke";
-
-async function resolveInput(
-  caps: Capabilities,
-  inputContent: LLMContent
-): Promise<string> {
-  const template = new Template(caps, inputContent);
-  const substituting = await template.substitute({}, async () => "");
-  if (!ok(substituting)) {
-    return substituting.$error;
-  }
-  return toText(substituting);
-}
-
-export type Inputs =
-  | {
-      context?: LLMContent[];
-      "p-location": LLMContent;
-    }
-  | WeatherInputs;
+export type WeatherInputs = {
+  location: string;
+};
 
 export type Outputs = {
   context: LLMContent[];
 };
 
 async function invoke(
-  inputs: Inputs,
+  inputs: WeatherInputs,
   caps: Capabilities
 ): Promise<Outcome<Outputs>> {
-  let location;
-  if ("context" in inputs) {
-    const last = inputs.context?.at(-1);
-    if (last) {
-      location = toText(last);
-    } else {
-      return err("Please provide a location");
-    }
-  } else if ("p-location" in inputs) {
-    location = await resolveInput(caps, inputs["p-location"]);
-  } else {
-    location = inputs.location;
-  }
-  location = (location || "").trim();
+  const location = (inputs.location || "").trim();
   if (!location) {
     return err("Please provide a location");
   }
   console.log("Location: " + location);
-  const weatherResult = await getWeather({ location }, caps);
-  if (!ok(weatherResult)) {
-    return weatherResult;
-  }
+  const executing = await executeTool<string>(caps, "get_weather", {
+    location,
+  });
+  if (!ok(executing)) return executing;
   return {
     context: [
       toLLMContent(
-        `Location: ${location}\n\n Weather information: ${toText(weatherResult)}`
+        `Location: ${location}\n\n Weather information: ${executing}`
       ),
     ],
   };
 }
 
-export type DescribeInputs = {
-  inputs: Inputs;
-  inputSchema: Schema;
-  asType: boolean;
-};
-
-async function describe({ asType: _, ...inputs }: DescribeInputs) {
-  const isTool = inputs && Object.keys(inputs).length === 1;
-  if (isTool) {
-    return getWeatherDescriber();
-  }
-  const hasWires = "context" in (inputs.inputSchema.properties || {});
-  const location: Schema["properties"] = hasWires
-    ? {}
-    : {
-        "p-location": {
-          type: "object",
-          title: "Location",
-          description:
-            "Please provide the location for which to get current weather",
-          behavior: [
-            "llm-content",
-            "config",
-            "hint-preview",
-            "hint-single-line",
-          ],
-          default: defaultLLMContent(),
-        },
-      };
+async function describe() {
   return {
     title: "Get Weather",
     description: "Searches weather information on Google Search.",
     inputSchema: {
       type: "object",
       properties: {
-        context: {
-          type: "array",
-          items: { type: "object", behavior: ["llm-content"] },
-          title: "Context in",
-          behavior: ["main-port"],
+        location: {
+          type: "string",
+          title: "Location",
+          description: "The name of the city",
         },
-        ...location,
       },
     } satisfies Schema,
     outputSchema: {
       type: "object",
       properties: {
-        context: {
-          type: "array",
-          items: { type: "object", behavior: ["llm-content"] },
-          title: "Context out",
-          behavior: ["hint-text", "main-port"],
+        weather: {
+          type: "object",
+          title: "Current Weather",
         },
       },
     } satisfies Schema,
